+++ frontend/admin/app/(app)/notifications/notifications-push/_components/NotificationForm.tsx
@@ -1,216 +1,223 @@
 import { Input, Textarea } from '@heroui/input';
-import { image, Select, SelectItem } from '@heroui/react';
-import { Star, School, BookMarked, CircleCheckBig } from 'lucide-react';
-import { useEffect, useState } from 'react';
-import { ImageUploader } from './NotificationImageUpload';
+import { Accordion, AccordionItem, Autocomplete, AutocompleteItem, Form } from '@heroui/react';
+import * as LucideIcons from 'lucide-react';
+import { Notification } from '@/types/notification';
+import ImageInput from '@/components/ui/imageInput';
+import { ComponentType } from 'react';
 
-type InformationInfoData = {
-  icon?: React.ElementType;
-  title: { en: string; th: string };
-  subtitle: { en: string; th: string };
-  body: { en: string; th: string };
-  redirect: { en: string; th: string; link: string };
-  imageUrl?: string;
-  imageFile?: File;
+type NotificationFormProps = {
+  notification: Notification;
+  onChange: React.Dispatch<React.SetStateAction<Notification>>;
+  onSubmit: () => void;
 };
 
-export const icons = [
-  {
-    id: 1,
-    name: 'Star',
-    icon: Star,
-  },
-  {
-    id: 2,
-    name: 'Book',
-    icon: BookMarked,
-  },
-  {
-    id: 3,
-    name: 'School',
-    icon: School,
-  },
-  {
-    id: 4,
-    name: 'Correct',
-    icon: CircleCheckBig,
-  },
-];
+export function NotificationForm({ notification, onChange, onSubmit }: NotificationFormProps) {
+  const allLucideIcons = Object.keys(LucideIcons)
+    .map(iconName => ({
+      value: iconName,
+      label: iconName,
+      icon: LucideIcons[iconName as keyof typeof LucideIcons] as ComponentType<LucideIcons.LucideProps>,
+    }))
 
+  const handleImageChange = (file: File) => {
+    onChange((prev) => ({
+      ...prev,
+      image: file,
+      imageURL: URL.createObjectURL(file),
+    }));
+  };
 
-type InformationinfoProps = {
-  onChange?: (data: InformationInfoData) => void;
-  resetSignal?: number;
-};
+  const handleImageCancel = () => {
+     onChange((prev) => ({
+      ...prev,
+      image: undefined,
+      imageURL: undefined,
+    }));
+  }
 
-export function InformationForm({ onChange, resetSignal }: InformationinfoProps) {
-  const [selected, setSelected] = useState<(typeof icons)[0] | undefined>(undefined);
-  const [imageUrl, setImageUrl] = useState<string | undefined>(undefined);
-  const [imageFile, setImageFile] = useState<File | undefined>(undefined);
-  const [title, setTitle] = useState({ en: '', th: '' });
-  const [subtitle, setSubtitle] = useState({ en: '', th: '' });
-  const [body, setBody] = useState({ en: '', th: '' });
-  const [redirect, setRedirect] = useState({ en: '', th: '', link: '' });
-
-  useEffect(() => {
-    const data = { icon: selected?.icon, title, subtitle, body, redirect, imageUrl, imageFile };
-    if (onChange) {
-      onChange(data);
-    }
-  }, [selected, title, subtitle, body, redirect, imageUrl, onChange]);
-
-  useEffect(() => {
-    setSelected(undefined);
-    setImageUrl(undefined);
-    setImageFile(undefined);
-    setTitle({ en: '', th: '' });
-    setSubtitle({ en: '', th: '' });
-    setBody({ en: '', th: '' });
-    setRedirect({ en: '', th: '', link: '' });
-  }, [resetSignal]);
+  const anyRedirectFilled = !!(
+    notification.redirectButton?.label.en?.trim() ||
+    notification.redirectButton?.label.th?.trim() ||
+    notification.redirectButton?.url?.trim()
+  );
 
   return (
-    <div className="flex flex-col gap-5">
-      <div className="flex flex-row justify-between">
-        <h1 className="text-xl font-bold  justify-center ">
-          Information Info
+    <Form 
+      id='notification-form' 
+      className="flex flex-col gap-4" 
+      onSubmit={(e) => {
+        e.preventDefault();
+        onSubmit();
+      }}
+    >
+      <div className="flex flex-row justify-between w-full">
+        <h1 className="text-xl font-bold">
+          Notification Form
         </h1>
-        <div className=" w-full max-w-[9rem] h-full">
-          <Select
-            className=" border border-gray-300 rounded-lg"
-            label="Select Icons"
-            size="sm"
-            selectedKeys={selected ? [selected.name] : []}
+        <div className=" w-full max-w-56 h-full">
+
+          <Autocomplete
+            isRequired
+            label="Notification Icon"
+            placeholder="Select Icon"
+            size='sm'
+            defaultItems={allLucideIcons}
             onSelectionChange={keys => {
-              const name = Array.from(keys)[0];
-              const item = icons.find(i => i.name === name);
-              setSelected(item);
+              const iconItem = allLucideIcons.find(icon => icon.value === keys);
+              onChange({ ...notification, icon: iconItem ? iconItem.value : '' });
             }}
-            renderValue={() =>
-              selected ? (
-                <div className="flex items-center gap-2">
-                  <selected.icon className="w-5 h-5" />
-                  <span>{selected.name}</span>
-                </div>
-              ) : null
-            }
           >
-            {icons.map(item => {
-              const Icon = item.icon;
+            {(icon) => {
+              const Icon = icon.icon;
               return (
-                <SelectItem
-                  key={item.name}
-                  startContent={<Icon className="w-5 h-5" />}
-                >
-                  {item.name}
-                </SelectItem>
+                <AutocompleteItem key={icon.value} startContent={<Icon className="w-5 h-5" />}>
+                  {icon.label}
+                </AutocompleteItem>
               );
-            })}
-          </Select>
+            }}
+          </Autocomplete>
+
         </div>
       </div>
 
-      {/* Title */}
-      <h1 className="text-lg font-bold"> Title </h1>
-      <div className="flex flex-row justify-between gap-5">
+      <div className="flex flex-row gap-4 w-full">
         <Input
-          label="English"
-          size="md"
-          type="text"
-          className="bg-white border border-gray-300 rounded-xl"
-          value={title.en}
-          onChange={e => setTitle({ ...title, en: e.target.value })}
+          isRequired
+          label="Title (English)"
+          placeholder="Enter title in English"
+          value={notification.title.en}
+          onChange={e => onChange({ ...notification, title: { ...notification.title, en: e.target.value } })}
         />
 
         <Input
-          label="Thai"
-          size="md"
-          type="text"
-          className="bg-white border border-gray-300 rounded-xl"
-          value={title.th}
-          onChange={e => setTitle({ ...title, th: e.target.value })}
+          isRequired
+          label="Title (Thai)"
+          placeholder="Enter title in Thai"
+          value={notification.title.th}
+          onChange={e => onChange({ ...notification, title: { ...notification.title, th: e.target.value } })}
         />
       </div>
 
-      {/* Subtitle */}
-      <h1 className="text-lg font-bold"> Subtitle</h1>
-      <div className="flex flex-row justify-between gap-5">
+      <div className="flex flex-row gap-4 w-full">
         <Input
-          label="English"
-          size="md"
-          type="text"
-          className="bg-white border border-gray-300 rounded-xl "
-          value={subtitle.en}
-          onChange={e => setSubtitle({ ...subtitle, en: e.target.value })}
+          isRequired
+          label="Subtitle (English)"
+          placeholder="Enter subtitle in English"
+          value={notification.subtitle.en}
+          onChange={e => onChange({ ...notification, subtitle: { ...notification.subtitle, en: e.target.value } })}
         />
 
         <Input
-          label="Thai"
-          size="md"
-          type="text"
-          className="bg-white border border-gray-300 rounded-xl "
-          value={subtitle.th}
-          onChange={e => setSubtitle({ ...subtitle, th: e.target.value })}
+          isRequired
+          label="Subtitle (Thai)"
+          placeholder="Enter subtitle in Thai"
+          value={notification.subtitle.th}
+          onChange={e => onChange({ ...notification, subtitle: { ...notification.subtitle, th: e.target.value } })}
         />
       </div>
 
-      <h1 className="text-lg font-bold"> Description </h1>
-      <div className="flex flex-row justify-between gap-5 items-stretch">
+      <div className="flex flex-row gap-4 items-stretch w-full">
         <div className="w-full h-full">
           <Textarea
-            label="English"
-            className="bg-white border border-gray-300 rounded-xl "
-            value={body.en}
-            onChange={e =>
-              setBody({ ...body, en: e.target.value })
-            }
+            isRequired
+            label="Description (English)"
+            placeholder="Enter description in English"
+            value={notification.body.en}
+            minRows={6}
+            onChange={e => onChange({ ...notification, body: { ...notification.body, en: e.target.value } })}
           />
         </div>
         <div className="w-full h-full">
           <Textarea
-            label="Thai"
-            className="bg-white border border-gray-300 rounded-xl "
-            value={body.th}
-            onChange={e =>
-              setBody({ ...body, th: e.target.value })
-            }
+            isRequired
+            label="Description (Thai)"
+            placeholder="Enter description in Thai"
+            value={notification.body.th}
+            minRows={6}
+            onChange={e => onChange({ ...notification, body: { ...notification.body, th: e.target.value } })}
           />
         </div>
       </div>
 
-      <h1 className="text-lg font-bold"> Redirect Button (Optional) </h1>
-      <div className="grid grid-cols-2 justify-between gap-5">
-        <Input
-          label="English"
-          size="md"
-          type="text"
-          className="bg-white border border-gray-300 rounded-xl "
-          value={redirect.en}
-          onChange={e => setRedirect({ ...redirect, en: e.target.value })}
-        />
+      <Accordion variant="shadow" selectionMode="multiple">
+        <AccordionItem 
+          key="1"
+          title="Redirect Button (Optional)"
+          startContent={
+            <LucideIcons.Link width={20} height={20} />
+          }
+          subtitle="Button to redirect users to a specific link."
+        >
+          <div className="flex flex-col gap-4 pb-3">
+            <div className='flex flex-row gap-4'>
+              <Input
+                isRequired={anyRedirectFilled}
+                label="Label (English)"
+                placeholder="Enter label in English"
+                value={notification.redirectButton?.label.en}
+                onChange={e => onChange({ 
+                  ...notification, 
+                  redirectButton: { 
+                    url: notification.redirectButton?.url ?? '', 
+                    label: { 
+                      th: notification.redirectButton?.label?.th ?? '', 
+                      en: e.target.value 
+                    } 
+                  } 
+                })} 
+              />
 
-        <Input
-          label="Thai"
-          size="md"
-          type="text"
-          className="bg-white border border-gray-300 rounded-xl "
-          value={redirect.th}
-          onChange={e => setRedirect({ ...redirect, th: e.target.value })}
-        />
-        <Input
-          label="Link"
-          size="md"
-          type="url"
-          className="bg-white border border-gray-300 rounded-xl "
-          value={redirect.link}
-          onChange={e => setRedirect({ ...redirect, link: e.target.value })}
+              <Input
+                isRequired={anyRedirectFilled}
+                label="Label (Thai)"
+                placeholder="Enter label in Thai"
+                value={notification.redirectButton?.label.th}
+                onChange={e => onChange({ 
+                  ...notification, 
+                  redirectButton: { 
+                    url: notification.redirectButton?.url ?? '', 
+                    label: { 
+                      en: notification.redirectButton?.label?.en ?? '', 
+                      th: e.target.value 
+                    } 
+                  } 
+                })}
+              />
+            </div>
 
-        />
-      </div>
-
-      <h1 className="text-lg font-bold"> Image (Optional) </h1>
-      <ImageUploader onChange={(file, url) => { setImageUrl(url); setImageFile(file ?? undefined) }}
-      resetSignal={resetSignal} />
-    </div>
+            <Input
+              isRequired={anyRedirectFilled}
+              label="URL Link"
+              placeholder="Enter URL link"
+              type="url"
+              value={notification.redirectButton?.url}
+              onChange={e => onChange({ 
+                ...notification, 
+                redirectButton: {
+                  label: notification.redirectButton?.label ?? { th: '', en: '' },
+                  url: e.target.value,
+                },
+              })}
+            />
+          </div>
+        </AccordionItem>
+        <AccordionItem
+          key="2"
+          title="Image (Optional)"
+          startContent={
+            <LucideIcons.Image width={20} height={20} />
+          }
+          subtitle="Image to display in notification."
+        >
+          <div className='pb-3 sm:w-full md:w-80'>
+            <ImageInput
+              title={'Notification Image'}
+              onChange={handleImageChange}
+              onCancel={handleImageCancel}
+            />
+          </div>
+        </AccordionItem>
+      </Accordion>
+    </Form>
   );
 }
